get requests

+++ java/simple-condo-app/complete/src/main/java/src/models/CondoDoorman.java
@@ -4,7 +4,7 @@ public class CondoDoorman extends Person{
 
     private final Double wage;
 
-    public CondoDoorman(Long id, String name, String cpf, Double wage) {
+    public CondoDoorman(int id, String name, String cpf, Double wage) {
         super(id, name, cpf);
         this.wage = wage;
     }
